made deletable tag view
